Move source fetch to deploy env task
This only needs to be run when environments are being deployed. All
other times there's a chance we'll fetch environments that don't exist
on disk and explode when operating on them.

Ideally this could be fetched one per run, but I can't determine an
elegant solution for this right now.