(r10k-#32) Support --verbose with no args
Cri uses the --verbose flag internally to display hidden commands. Since
a number of commands in 1.0 were deprecated, there has been this
alluring list of commands that are hidden, but using --verbose blows up
with the following:

    r10k: option requires an argument -- verbose

The --verbose flag previously required an argument rather than
arbitrarily pick a log level. To make --verbose work for displaying the
help, an arbitrary log level of 'info' was picked. This way, using
--help --verbose does what you would expect, rather than fail in a
rather obtuse manner.