(r10k-#32) Support --verbose on commands with no action
A number of commands have no implementation and mainly serve to act as
namespaces. When these commands are directly called, they return the
help to make use a bit more intuitive. Cri::Command#help takes an
optional :verbose option, and directly invoking it doesn't magically
transfer the argument from the command `run` block to the #help method.

To make sure that --verbose behaves as expected with these
implementationless commands, opts[:verbose] needs to be passed
explicitly.