(r10k-#48) Default environment prefixing to false
Defaulting environment prefixing to true with multiple sources are
present can have unexpected outcomes if users are already using multiple
sources in different directories. While it could be possible to only add
prefixing to multiple sources with the same basedir, it's nontrivial and
right now there's a backwards incompatible change in a released version
that needs to be rolled back.

This commit removes the prefix detection so that prefixing must be
explicitly enabled.