[module/forge] Delegate to metadata, simplify tests
This commit changes the forge module to handle most of the metadata
interaction to the dedicated class, simplifying the actual work that
needs to be done inside of the forge module. It also guts and overhauls
a lot of the test suite and fixes bugs that were uncovered while
changing out other functionality.