(GH-51) [module/forge] Always expand expected_version
Forge modules can have a version of `:latest`, which will set the
expected version based on the latest version of a module on the forge.
However this means that the variable always needs to be expanded if it's
latest, and there were instances where the value was not expanded.
This commit resolves the issue by wrapping all uses of `@expected_version` in
a helper method that will lazily expand the value on demand.