[git] Collapse #resolvable? function
The original implementation of #resolvable? was determining both if the
ref was resolvable as well as up to date. Determining the resolvability
of a ref is simple and behaves the same across all git primitives, while
determining freshness varies between immutable refs (tags, commits) and
mutable refs (heads). Since #fetch? was split out for the second case,
we can make that vary between different object types and use the same
code for determining resolvability.