[util/subprocess] Add initial childprocess based subcommand
As it turns out, 'systemu' is slow. Very slow. Like 50x slower than
using %x[], roughly. Given a crude benchmark:

%x             0.000000   0.040000   0.240000 (  0.289449)
system         0.010000   0.030000   0.180000 (  0.325935)
childprocess   0.010000   0.030000   0.140000 (  0.196935)
systemu        0.080000   0.060000  18.480000 ( 18.538550)

Since r10k shells out to a lot of commands, this butchers performance
for longer runs.

This commit adds a basic childprocess based command runner. It should be
more flexible and reusable than R10K::Execution, and both can be
installed in parallel as things are moved over.

The major caveat is that child processes grab the TTY while running so
sending a SIGINT will be send to the child process. This has been a
problem before but for some reason it's pretty badly exacerbated by
switching to childprocess. At some point we need to ensure that child
processes have a different sid.