[gemspec] (GH-82) include all files in gemspec
The gemspec was only pulling in files in the bin, lib, and spec dirs,
which works fine for packaging as a gem but makes it more challenging to
repackage r10k in other format. This commit rectifies this by packaging
all files managed by git, which is what a number of other ruby projects
do.