(GH-90) Prevent name collisions with multiple sources
Previously, when environments were being interacted with, a lookup table
was being created of names and associated environments for determining
what environments should be acted upon. However this didn't take into
account multiple sources where multiple environments could coexist with
the same name. In this case the duplicate environments would be silently
ignored.

This is resolved by getting rid of the lookup table and just scanning a
list of environments with the given names. It's a bit slower can could
theoretically take a fair amount of time, but that would require a
situation with 1k+ environments. For most normal operations
this will be very fast, and won't have the same collision issues.