[cli/deploy] (GH-112) per-environment module update
When updating a module within a deployment, it's common to want to
update that module in a single environment. However the previous
behavior of `r10k deploy module <module>` would deploy the module in all
environments.

This commit adds a '--environment' option to `r10k deploy module`. It
allows a single environment to be specified and all module updates will
occur in that environment.