[source] Transpose order of name and basedir in #initialize
The order of arguments in sources and environments was transposed in
(#initialize), where sources had (basedir, name, opts) and environments
had (name, basedir, dirname, opts). This made the interactions between
the two classes a bit strange made it easy to reverse argument order.
This commit switches the order so that the general layout is
(descriptor, file path arguments, opts).