[git/working_dir] (GH-163) delete tags when changing remotes
Tags are not namespaced and are not overwritten by default, so when a
git repository remote is swapped out the tags from the old remote may
linger and could collide with the new tags. This commit changes the
remote update logic to purge all local tags when the remotes are swapped
out.