[spec/system] (maint) Remove rspec-system tests
The rspec-system tests have been broken for a while, rspec-system has
been end of lifed, and it doesn't look like these tests are going to
receive any love any time soon. Porting the tests to beaker-rspec or
some other testing framework will probably mean rewriting them almost
entirely and module installation tests need to be completely revised, so
all of this is providing zero value. This commit removes all of the
rspec-system tests and supplementary code; if we need this back we can
use the Git history to perform that resurrection.