[puppetfile] (GH-186) Rescue Puppetfile parse errors
When a Puppetfile with invalid syntax is parsed it raises a SyntaxError,
and the `r10k puppetfile check` code was not specifically handling that.
Thus when checking an invalid file r10k was actually crashing and not
gracefully handling the error.

This commit cleans up the puppetfile check logic and adds a message on
valid parse to indicate success.