[module] Allow SVN modules to have owner/modname titles
Right now modules are selected based on the derived class providing a
`.implement?` method that indicates whether it can install the given
module. This is a rather loose model and ultimately it should be
replaced with a KeyedFactory instance. In addition this has strange
ordering behaviors, where SVN modules could not have an
'owner/modulename' title because the Forge module would get to it first.

This commit works around that deficiency by requiring the Forge module
afterwards so that the SVN can register itself first and implement those
modules.