[svn/working_dir] Run 'svn checkout' with --quiet
Right now we're relying on commands to return less than 64k of data on
stderr and stdout because we just queue everything in pipes. If we fill
said pipes we hang, and SVN is very chatty and will cause that hang. To
resolve this for now this adds '-q' to checkout invocations to bypass
this bug.