[module_repository/forge] Force json_pure on 1.8.7
When multi_json is running under Ruby 1.8.7, the json gem isn't present,
but the json_pure gem is present, the multi_json library may mistake the
json_pure gem as the json gem, try to load the latter, and fail. r10k
specifically pulls in json_pure for 1.8.7 compatibility and our current
use of JSON parsing is pretty minimal so we don't need blazing speed. To
simplify the use of JSON and ensure consistent behavior we force use of
json_pure on 1.8.7.