Add 'strip_branchname_prefix' option for git
Gitolite supports 'personal branches', which are a convenient way to
allow a large group of developers push access to any number of branches
within their own namespace. However it requires a prefix on the branch
name (e.g. "personal/") which just adds noise to the name of the Puppet
environment. This commit provides a 'strip_branchname_prefix' option, to
allow you to strip a fixed prefix from a branch name before using it.