[action/deploy/environment] (GH-260) Handle \W in environment names
Version 1.4.0 broke handling of environment names containing \W. When
determining which branch to check out, it compares the unnormalized
environment name(s) passed in on the command line to the normalized
environment name returned by the backend. This normalizes the names
passed in on the command line before the comparison happens.