[git/cache] (RK-16) Use BareRepository from Git cache
The previous implementation of the Git caching code intermixed caching
logic and Git commands which made things hard to test and reason about.
This commit removes all of the Git implementation logic from the cache
class and reuses the BareRepository class, so the Cache only has to deal
with, well, caching.