[git/working_repository] (RK-16) Implement Git non-bare repo
The WorkingDir class implemented interactions with non-bare git
repositories, but it blended handling a non-bare repository, performing
sparse clones, and enforcing the state of a given repository. This
commit implements a class for just interacting with a non-bare
repository; the concerns of sparse clones and state enforcement will be
handled by separate classes.