[working_repository] (RK-16) Use #checkout method after initial clone
WorkingRepository instances could clone a repository and check out an
initial ref, but `git clone -b` can only take a branch or tag and not a
commit. In order to resolve this consistency and make
branches/tags/commit be equivalent this commit uses #checkout which
treats the three as equivalent.