[git/stateful_repository] (RK-16) Treat unsynced branches as outdated
Git branches are a bit strange in comparison to Git commit SHAs and tags
because they are mutable. If we've been given a Git ref that's a branch,
we have to make sure that the branch itself has been fetched so that we
actually track upstream changes to the branch.