[util/exec_env] (RK-17) Add safe env modifiers
In order to expand shell commands we need to be able to read environment
variables, and in order to test that code we need to be able to modify
the environment. This commit adds helpers to modify the current
environment within a specified scope so that changes don't affect the
entire running process.