[git] (RK-16) Use def_delegators to forward method calls
Before this commit, the Git environment and module had moved most of
their behavior to the backing repo class, but did that delegation by
calling the delegated method. A number of other places were using
`def_delegators` to provide this behavior instead. While it's a small
thing, the environment and module classes were inconsistent.

This commit updates the Git environment and module classes to use
`def_delegator` for methods that are purely delegated to the backing
repo class.