[Gemfile] Add rugged gem in 'extra' group
As part of adding a libgit2 based Git implementation we need to add
rugged as a dependency, but is not a hard requirement for r10k. This
commit adds rugged as an optional dependency. Rugged does not supported
on Ruby 1.8.7 so is excluded on that Ruby version; on 1.9 and greater it
will be included unless bundler is run with `--without extra`.