[git/cache] Initialize instance cache in class body
Without this commit, it's not possible to subclass R10K::Git::Cache and
have separate instance caches because of how the cache was initialized.
The cache was lazily constructed inside of the getter method, which had
the interesting side effect that subclasses would always share this
class. This means that different Git implementations might intermix,
which isn't desirable.

This commit moves the InstanceCache instantiation to outside of the
getter method, so that subclasses can provide separate instance caches.