[settings/mixin] Look up settings in superclasses
When a class with settings is subclassed, the settings applicable to
the superclass are generally applicable to the subclass. In order to
link up the  super/subclass relationship, this commit uses the
Class#inherited hook to manually create a settings instance in the
subclass that pulls settings from the superclass.