[spec] Generate test coverage reports for all files in lib
The previous simplecov path filter would only report code coverage for
files in lib/r10k; with the inclusion of files outside of that directory
this pattern excludes files we may need to test and include in test
coverage reports. This commit updates the filter to include all files in
the libdir of the project as part of the code coverage report.