(RK-27) Move request to find latest release to v3 forge api.
Before this r10k made a request to the deprecated unversioned forge
API to find the latest release of a module. After this it will get
that information from the v3 API. If a user has forge.puppetlabs.com
as the forge in a Puppetfile it will be changes to
forgeapi.puppetlabs.com and the user will be warned.