[git/stateful_repository] Only sync Git cache when necessary
Updating the cached Git repo generally means a network call, which is an
enormous performance bottleneck. There are a few cases where we can skip
this call entirely - specifically when a repo is pinned to a commit or
tag. Tags and commits are immutable, so if they can be resolved against
the cached git repo then we can skip the network call. This commit adds
logic to the stateful repository to do this check, and opportunistically
skip slow network calls.