[git/stateful_repository] don't fetch thin repo on sync
The git alternates mechanism allows a git repository to borrow objects
from another repository, which allows for a great deal of time and space
savings. One of the implications of this is that if only commits are
used and checked out, we never need to actually fetch any refs or
objects from the remote repo. This commit removes the #fetch call when
updating an existing repo and relys on the borrowing logic to check out
the correct object.