[rugged] (RK-17) Mark rugged_repo as lazily evaluated
The rugged_repo instance variable needs to be lazily generated and
accessed, as if it's created with a directory that's not a git
repository it will fail. To make sure we don't accidentally access the
variable instead of with the appropriate helper method, this commit
prefixes the instance variable with _ to indicate that it's lazily
generated.