[git] (RK-17) Add Git provider detection
We will have multiple providers for Git functionality; the current
shellgit code and a new libgit2 based implementation. In order to swap
the two implementations out we can't directly instantiate classes but
need to look up the current implementation and create instances from
that. This commit adds the abstraction needed to create instances from
the currently selected Git implementation.