[deployment] (RK-4) Raise meaningful error on missing sources
If the 'sources' key in r10k.yaml was left unset, was misspelled, or was
empty, r10k try to blindly iterate through it as a hash and would
subsequently raise an error. This commit updates the source enumeration
code to check for these error case and raise a meaningful error.