[util/symbolize_keys] Extract #symbolize_keys! method
The existing #symbolize_keys! method needed callers to monkey patch
themselves in order to use the method, which was less than ideal. doing
monkey patching like this at runtime invalidates the Ruby internal
method cache which can cause a non-trivial performance hit, and the
monkey patching was done selectively in the first place.

This commit defines a standalone `symbolize_keys!` method that can be
invoked without monkey patching and replaces the old monkey patched
method with a call to the new implementation.