[util/symbolize_keys] Recursively symbolize hashes
When a data structure is read from a config file, be it YAML or JSON, it
may have nested hashes that all have string keys. This commit updates
.symbolize_keys! to optionally recurse into nested hashes.

Note that this will not recurse into arrays containing hashes; R10K
doesn't currently use arrays in config files so that behavior has been
omitted till it's needed.