[source] remove unused Purgeable methods and mixin.
Commit a96f98b6fee01b37fc7dc3308f13c36836677636 moved the
responsibililty of purging directories from R10K::Source instances to
R10K::Util::Basedir, thus making the R10K::Util::Purgeable code in the
SVN and Git sources unused (with the exception of #desired_contents).
This commit removes the dead code and updates the documentation to
indicate the actual state of the world.