[git] (RK-33) Stay failed on invalid provider
The original implementation of the provider code would disallow setting
an invalid provider, but it would be possible for the provider selection
to use a fallback value if called again. This meant that in certain
conditions it would be possible for an invalid value to be ignored and
the default provider silently used in its place.

This commit adds a hard failure mode to the provider selector. If an
invalid provider is selected, the class will effective "lock" itself and
raise an error so that re-querying for a provider will fail until the
class is explicitly reset.