[doc] Update dynamic env basedir docs about collisions
GH-202 updated the R10K::Deployment class to treat duplicate
environments as a critical error; if two sources create environments
that try to manage the same path r10k will fail hard. This commit
updates the source/basedir docs to reflect that an error will be raised
instead of ignoring the issue.