[source/svn] Remove `remote` from BranchName class
The BranchName class was meant to deal with branch validation and
manipulation, but the SVN implementation stuffed in extra information
about the SVN path as well. This was done for convenience but polluted
the class with unrelated information.

This commit pulls out the remote field from the BranchName class so that
it behaves in a more generic fashion. The calling code packs a
BranchName object and remote path in an array to group the data together
while avoiding the data pollution.