[logging] Eagerly initialize formatter/outputter
Trying to lazily generate the logr4 formatter and outputter is a lot of
work and requires using some fairly sketchy metaprogramming, and in the
end it's ultimately unnecessary. Because some basic logger configuration
must always be present, this extracts the formatter/outputter
configuration to occur when the logger class is loaded.