[rugged] Add log messages for git operations
The ShellGit provider takes a very simple approach to logging in that
all actions executed shell out, and the act of shelling out is logged.
This means that it's trivial to inspect the operations of the ShellGit
provider without having to add explicit log statements.

In contrast, the rugged provider does not have this convenient join
point for all logging; without explicit log messages it's impossible to
inspect what the libgit provider is doing.

This commit adds an initial pass of log messages to the rugged code. It
invokes the logger methods with a block (`#debug { ... }`) So avoid
building the log message when that log level is disabled.