[logging] Use log4r logger repository
The intention of the `#logger` method was to provide a single logger per
class to add context for the log messages, but reuse a single instance
per class to cut down on object overhead. However this does not happen
automatically, and the existing code was creating many more logger
instances than what was intended, which had a nontrivial overhead.

When a log4r logger is created, it's automatically stored by log4r in an
internal repository for later use. The logger has to be explicitly
created as requesting a logger with a given name will return nil if that
logger hasn't been created, but can be retrieved afterwards. This commit
modifies the `#logger` method to try to retrieve the class logger from
the log4r repository first, and if that fails generate a new logger.