[logging] Don't precompute log level mapping
Log levels are selected by the name of the level (eg "debug2") but Log4r stores
log levels as integers. Log4r stores the separate levels under
constants, so to resolve the name of a log level to the numeric level we
need to look up a constant value. The original code was precomputing and
caching this information, but given that log levels are generally set
once for the life of the process this is a premature optimization. This
commit removes the extra logic.