[logging] Centralize "default" log level handling
If a user specifies '--verbose' without a level, Cri will set the
verbosity as "true:TrueClass" and pass this along. This was originally
handled in the CLI code but it was far removed from logging, and if the
log level was invalid an exception would be thrown during argument
parsing.

This commit adds handling of 'true' as a log level and removes the
extra code path.