(CODEMGMT-129) Git Provider Switch for Integration Tests
There is a new option in the configuration file for r10k that allows the user
to change the Git provider. The integration tests have been altered to allow
for dynamic switching between providers. The "test_run_scripts" have been
updated to all the tester to choose which Git provider to use for testing.