[feature] (RK-73) Add logging for feature detection
Right now determining if a feature is available is entirely opaque; if a
feature failed to load there's no indication why.  For instance, the
rugged provider tries to load the 'rugged' gem - if that fails for any
reason (such as a missing shared object) the error is masked and r10k
will pretend that the rugged library is not present.

This commit adds log messages so that if a feature is unavailable the
failure can be divined by checking the debug level logging.