[git/stateful_repository] (RK-86) Correctly sync branches
R10k 1.5.0 added an optimization for avoiding syncing the Git cache
unless necessary, but accidentally treated all refs as commits, causing
branches to not be updated. This only affected Git modules as
environments/sources would always force a sync. This commit resolves the
issue by checking the type of the ref, not the type of the resolved sha,
when determining if the cache needs to be updated.