[settings/definition] Pass definition object to default block
The lambda that resolves the default value of a setting may require
information about the definition that it's adding a default for, or may
depend on another setting. This commit passes the definition into the
lambda so that this information can be reached. Since lambdas do arity
checking, default lambdas that do not needs this value should have an
arity of 1 but can ignore the passed value.