(RK-113) Remove deprecated Task classes and namespace
The R10K::Task namespace was the original mechanism for organizing the
r10k library code into a form that could be called as an application,
but was overly complex and hard to follow. It has been replaced by the
R10K::Action namespace and classes. This commit removes the Task classes
and namespace.