(RK-121) Provide error context for nonexistent forge releases
Without this change, if a user tries to download a module on the Forge
that doesn't exist they will receive an error indicating that some HTTP
request returned a HTTP 404 status, but nothing more. This doesn't
indicate what actually happened - that the requested module release
doesn't exist and can't be downloaded. This commit catches the general
404 error and indicates what actually happened.