[runner/windows] (RK-143) Use arg vector for windows commands
When run on Windows, r10k uses Open3.capture3 to execute commands, but
was joining the argv into a single string. Open3.capture3 calls .popen,
which in turn calls Process.spawn, which in turn accepts an argument
vector, so this commit changes the .capture3 call to pass the argv so
that it's called without modification.