(RK-156) Probe for pe_license library via require
If the pe_license module is installed into the ruby library path via
something like rpm or deb and is not installed into the rubygems library
path, it will not be locatable via `Gem::Specification.find_all_by_name`.
This commit switches the library detection to use the r10k feature
system, which handles detection by actually requiring the library. While
requiring the library as part of detection is a fairly blunt approach
and has side effects, it's a more reliable detection method (and if we're
checking to see if the library is present, we're going to want to require it
anyways).