Mention potential issues with rugged
There's a number of issues that can trip people over when using Rugged.
This should document a number of potential problems that people might
discover.

We mention what libraries are needed to be installed for libgit2 to
compiling support for https and ssh transports.

We mention that they can test their rugged installation using irb, or
that r10k 2.0 and later issues warnings if they're missing

We also mention an issue with Debian-based distributions and libssh2,
and provide an example error, so that users can find the documentation
by searching for that error.