[action] (RK-120) Add base action class
This commit adds a base action class and updates the action classes to
inherit it. This gives us a common interface for creating actions so
that we can pass common values (such as application settings) to all
classes on initialization without having to duplicate said functionality
into each individual class.