(RK-90) [module/forge] Ensure module metadata is set
In certain cases, trying to determine the status of a Forge module would
try to access metadata that had not been set. This usually happened when
the Forge object was created before the module had been created, and #status
was called after the module was created. In this case the metadata would
not be properly set and checking the status of the module would raise an
exception.

This commit properly updates the metadata instance variable so that r10k
doesn't fail hard in the described scenario.