(RK-21) Treat non-existent environments as error case
When r10k deploys a specific set of environments from the command line,
it would silently ignore environments that did not exist and return with
an exit code of 0. This commit ensures that when environments are listed
on the command line r10k will make sure they all exist, and when
environments are missing it will log an error and return a non-zero exit
code.