(RK-175) [rugged] Don't check out unresolvable refs
The rugged working repository always tried to resolve a Git ref to a SHA
when checking out a ref, but while the shellgit implementation could
rely on the underlying git binary to report meaningful errors the rugged
implementation had no such fallback check. This could lead to r10k
trying to check out 'nil' when an unresolvable ref was given, which
would raise an ArgumentError. This commit adds an explicit check for
unresolvable refs and raises a meaningful error instead of an
ArgumentError.