[git] (RK-174) Always check alternates file is up to date
When r10k interacts with thin repositories, it can't rely on the
alternates file having the correct contents because the cachedir setting
may have changed, and repositories that it's operating against may still
reference the old location. This can cause issues such as the wrong
cached repository being fetched which will mean that the thin repository
is effectively missing objects, which is very messy.

This commit updates the rugged and shellgit thin repo implementations to
always check and conditionally update the alternates file before trying
to read the contents of the file.

This may have performance ramifications when interacting with
repositories in a read only manner, but r10k is generally write heavy
and refactoring this to be more efficient is not worth the effort, since
in the long run we'll need to abandon the thin repo method in order to
support subdirectory checkouts.