[runner] (RK-169) Print validation errors
The settings framework that was added would collect all validation
failures and include them in the validation exception, but there was no
code handling that exception. Because of this, if validation failed the
error message would be "Validation failures for global" which is
entirely useless.

This commit adds a method to the validation error class that will format
the nested validation and support for catching validation errors and
printing out all validation errors.