(maint) Duplicate parent values during lookup
When a container looked up a value and found it in the parent container,
it would cache it and return it. However this did not account for cases
where the looked up value was changed by the caller, which would cause
the parent value to be changed as well. This commit dups values looked
up in the parent to limit the scope of changes.