(RK-220) Fail early when git private key is unreadable
When r10k was using rugged and Git SSH remotes with pubkey
authentication, r10k didn't check if the specified SSH key was readable
and let rugged do that checking. However rugged's error messaging around
this was less than clear when the private key was missing or unreadable
which users found confusing. This commit remedies the error messaging by
doing the check inside of r10k and failing early with a more readable
message.