(RK-243) Refuse to generate rugged credentials more than 50 times.
The new behavior as of libgit2/rugged 0.24.0 is to continue to call the
credentials callback until either it raises an error or the server gives
up. Since neither was happening for HTTP sources, invalid credentials were
leading to an infinite retry loop. SSH sources already seemed to have a
client side abort at around 50 attempts.