(CODEMGMT-702) Add code comments
This commit is to update the code comments.

Basically, the issue is that rugged/libgit2 won't update (at least) the
execute bit on a checkout.  And it won't update it on a reset --hard to
the same SHA, apparently.

sync is called from deploy::environment with force = true, and reset
--hard has the same effect on the working directory in the end, so the
checkout isn't necessary.