(CODEMGMT-1064) Test pe-r10k project before promoting it.
Prior to this commit the acceptance task would only target a previously packaged version of pe-r10k.  This addition allows for the installation of a dev version of pe-r10k over the install of pe. The acceptance task has been modified to allow for this additional component focused test, vs. a pre-suite test targeting a PEZ build. The ci:test:component task will need to be called by a component configuration to be added to ci-job-configs.