(CODEMGMT-1097) Remove assertion during duplicate module test
The integration test which tests what happens when duplicate module
names are found in an environment expects that one of the modules gets
deployed once but the actual behavior (as an outcome of the fix) leads
to an error and nonzero exit code from the r10k deploy. This commit
removes the erroneous test assertion.