(RK-297) Workaround rugged/r10k issue in r10k_utils.rb
Cleaning up the r10k cache between test runs exposed a bug in r10k's
rugged provider. If the cache has been removed and the control repo
history has been rewritten, deploys will fail with errors that mention
object not found. In order to work around this bug, this commit adds a
call to remove the git history of codedir as part of the clean_up_r10k
method. This allows deploys with the rugged provider to proceed
successfully.