Add boolean flag support to setopts
This commit adds support to setopts for :boolean flags. Specifying a
boolean flag means the flag when passed directly, --flag, is boolean
true. When the flag is denoted boolean, it will also be possible to pass
--no-flag, which will set the flag to false.

In the R10K::CLI module, if you want both true and false to be valid
flags you must reference both of them. (If you only want one value to be
passable, provide only the true or false version of the flag, as you
like). Example:

    module R10K::CLI
      module Example
        def self.command
          @cmd ||= Cri::Command.define do
            name    'example'

            flag nil, :'flag',    'Lets the user set flag to true'
            flag nil, :'no-flag', 'Lets the user set flag to false'

            run { |opts, args, cmd| exit 0 }
          end
        end
      end
    end

In allowed_initialize_opts, it is only necessary to define the flag
name once. E.g. for a flag "flag":

    def allowed_initialize_opts
      super.merge(:flag => :boolean)
    end