(CODEMGMT-1177) Add gitdir parameter for repos
This is necessary in order to later allow repos to be created that use a
gitdir besides the default .git, and to make checkout/cloning behaviors
more consistent across operations - especially those involving the git
dir.

A specific operation this enables is removal of the gitdir, followed by
creation of a new gitdir in its place. A working tree clone would force
removal or emptying of the entire working directory to swap out the
gitdir. Working with the gitdir separately opens up this use case.

The implementation choice of just ALWAYS working as a bare repo clone
followed by a checkout is to reduce codepaths. The work done to create a
working repository using gitdir=.git is the same work done to create a
working repository using gitdir=.git-altdir