(CODEMGMT-1152) Add expect_dirty arg for git
When a basemod is in use, it will inherently "dirty" the root of the
environment. Therefore, it will be useful to be able to tell basemods
and the control-repo both to expect some dirtiness, and not issue
warnings about it.

This commit implements the ability to expect dirtiness. Later commits
will use this capability.