(CODEMGMT-1152) Fix dirty warns using basemod
If a Puppetfile for an environment expresses a basemod, it is normal and
expected that the environment dir will be "dirty" as far as VCS is
concerned. Therefore, don't print warning messages about it.

Note that as part of this change, a shortcircuit/idempotency measure is
added to Puppetfile#load.

Further, a basemod being deployed into a directory already containing at
least a Puppetfile will also be dirty as far as VCS is concerned. This
is expected. Do not print dirty warnings for basemods either.